RT-NIDS-Malicous-URL-Detection-master/requirements.txt
absl-py==0.7.1
asn1crypto==0.24.0
astor==0.8.0
astroid==2.2.5
attrs==19.1.0
autopep8==1.4.4
backcall==0.1.0
bleach==3.1.0
bokeh==1.2.0
CacheControl==0.12.5
cachetools==3.1.1
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
configparser==3.7.4
cryptography==2.7
cssselect==1.0.3
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
dicttoxml==1.7.4
Django==2.2.3
django-memoize==2.2.0
djangorestframework==3.9.4
docutils==0.15.2
dpkt==1.9.2
entrypoints==0.3
firebase-admin==2.17.0
future==0.17.1
gast==0.2.2
geocoder==1.38.1
geoip2==2.9.0
google-api-core==1.13.0
google-api-python-client==1.7.9
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-cloud-core==1.0.2
google-cloud-firestore==1.2.0
google-cloud-storage==1.16.1
google-pasta==0.1.7
google-resumable-media==0.3.2
googleapis-common-protos==1.6.0
graphviz==0.11.1
grpcio==1.21.1
h5py==2.9.0
httplib2==0.13.0
idna==2.8
imbalanced-learn==0.5.0
ip2geotools==0.1.5
IP2Location==8.1.0
ipaddress==1.0.22
ipykernel==5.1.1
ipython==7.6.1
ipython-genutils==0.2.0
ipywidgets==7.5.0
isort==4.3.21
jedi==0.14.0
Jinja2==2.10.1
joblib==0.13.2
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.5.0
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.1
lightgbm==2.2.3
Logbook==1.4.3
lxml==4.3.4
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.0
maxminddb==1.4.1
mccabe==0.6.1
mistune==0.8.4
msgpack==0.6.1
nbconvert==5.5.0
nbformat==4.4.0
ndg-httpsclient==0.5.1
notebook==5.7.8
numpy==1.16.4
packaging==19.0
pandas==0.24.2
pandocfilters==1.4.2
parso==0.5.0
pexpect==4.7.0
pickleshare==0.7.5
pika==1.0.1
Pillow==6.1.0
pip-review==1.0
pkginfo==1.5.0.1
plotly==3.10.0
prometheus-client==0.7.1
prompt-toolkit==2.0.9
protobuf==3.8.0
psycopg2==2.8.3
ptyprocess==0.6.0
pusher==2.1.3
pusher-push-notifications==1.1.0
py==1.8.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycodestyle==2.5.0
pycparser==2.19
pydot==1.4.1
pydotplus==2.0.2
Pygments==2.4.2
PyJWT==1.7.1
pylint==2.3.1
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.0
pyquery==1.4.0
pyrsistent==0.15.2
pyshark==0.4.2.3
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1.1
pyzmq==18.0.2
qtconsole==4.5.1
ratelim==0.1.6
readme-renderer==24.0
requests==2.22.0
requests-file==1.4.3
requests-toolbelt==0.9.1
retrying==1.3.3
rsa==4.0
scapy==2.4.2
scapy-http==1.8.2
scikit-learn==0.21.2
scipy==1.3.0
seaborn==0.9.0
selenium==3.141.0
Send2Trash==1.5.0
six==1.12.0
sklearn==0.0
sqlparse==0.3.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
tldextract==2.2.1
tornado==6.0.3
tqdm==4.32.2
traitlets==4.3.2
twine==1.13.0
typed-ast==1.4.0
typing==3.7.4
uritemplate==3.0.0
urllib3==1.25.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.4
whois==0.8
widgetsnbextension==3.5.0
wrapt==1.11.2
